fix: CI sign-compare error, Copilot review fixes, add turn_ to symbol allowlist
- Fix signed/unsigned comparison in turn_unwrap_channel_data() (nano_turn.c:743)
- Add bounds check in test helper add_candidate() to prevent buffer overflow
- Extend test_trickle_max_candidates to verify overflow rejection via public API
- Exclude network-dependent interop_turn from CI ctest (-LE network)
- Add turn_ prefix to symbol allowlist in ci-check.sh and ci.yml
- Auto-fix clang-format on NANORTC_CONFIG_DEFAULT macro

+++ src/nano_turn.c
@@ -740,7 +740,7 @@ int turn_unwrap_channel_data(const uint8_t *data, size_t len, uint16_t *channel,
     if (*channel < 0x4000 || *channel > 0x7FFF) {
         return NANORTC_ERR_PROTOCOL;
     }
-    if (4 + plen > len) {
+    if ((size_t)4 + plen > len) {
         return NANORTC_ERR_PARSE;
     }
